+++ app/app.vue
@@ -4,6 +4,7 @@
         <NuxtPage />
         <ToastHolder />
         <OnboardingOverlay />
+        <PageScrollbar />
     </div>
 </template>
 <script setup lang="ts">